Move Batch datatype to its own file
